Create phildini-bookwyrm-troubleshooting.md (#50)
Added this troubleshooting document that's mostly just notes to myself. Maybe it's useful to find places to add some of this in the project-level docs, but some of it may also already be covered, and it's definitely all stuff that should be obvious, but since I keep forgetting to doublecheck these things, let's make a list!